+++ packages/orq-rc/RELEASES.md
@@ -8069,3 +8069,13 @@ Based on:
 - [typescript v4.7.0-rc.50] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.50] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.50 - packages/orq-rc
+
+## 2026-04-08 06:35:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.51] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.51] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.51 - packages/orq-rc